Organize DataImport models and readers
