+++ src/DataImport/Models/StudentVleCsv.cs
@@ -1,13 +1,13 @@
 using CsvHelper.Configuration.Attributes;
 
-namespace OuladEtlEda.DataImport;
+namespace OuladEtlEda.DataImport.Models;
 
-public class VleCsv
+public class StudentVleCsv
 {
     [Name("id_site")] public int IdSite { get; set; }
+    [Name("id_student")] public int IdStudent { get; set; }
     [Name("code_module")] public string CodeModule { get; set; } = null!;
     [Name("code_presentation")] public string CodePresentation { get; set; } = null!;
-    [Name("activity_type")] public string? ActivityType { get; set; }
-    [Name("week_from")] public int? WeekFrom { get; set; }
-    [Name("week_to")] public int? WeekTo { get; set; }
+    [Name("date")] public int? Date { get; set; }
+    [Name("sum_click")] public int SumClick { get; set; }
 }
