+++ src/DataImport/Readers/CsvAssessmentReader.cs
@@ -1,6 +1,7 @@
+using OuladEtlEda.DataImport.Models;
 using CsvHelper.Configuration;
 
-namespace OuladEtlEda.DataImport;
+namespace OuladEtlEda.DataImport.Readers;
 
 public class CsvAssessmentReader : CsvReaderBase<AssessmentCsv>
 {
